requirements.txt
ultralytics
pyside6
opencv-python
paho-mqtt
Pillow
numpy
